+++ CHANGELOG.md
@@ -1,6 +1,9 @@
 # Changelog
 
 ## [Unreleased]
+### Added
+- Opaque relay state
+- Logging of AuthnRequest and Response
 
 ## [0.18.0] - 2018-09-12
 ### Removed